Expose EdgeDBDecodeDateTime() and EdgeDBToTimestamp() functions
These allow to implement a stricter date/time parsing in EdgeDB.
Contrary to Postgres, it's an error to supply a timezone to a "naive"
date/time type, such as "timestamp"; it's also an error to not to
supply a timezone to a timezone-aware type, such as "timezonetz".

The exported functions allow us to implement strict flavors of
time_in(), date_in(), timestamp_in(), timestamptz_in(), to_timestamp(),
and to_timestamptz().