Avoid table rewrite when adding nullable columns of constrained domains (#3)
Postgres has an optimization whereby it skips a full table rewrite if a
default was specified or the column is nullable, however this
optimization is skipped if a domain has constraints defined on it,
because `CHECK` constraints get invoked on NULL values and there is no
way for Postgres to know how that NULL value affects the constraint
without evaluating it.  Unfortunatly, there's no direct way of
evaluating domain constraints from that part of `ALTER TABLE`
processing.  However, EdgeQL does not support checking for an empty
value in scalar constraints, so we can safely skip the table rewrite on
nullable columns even if there are constraints (but still keep it for
domains marked as `NOT NULL` just in case).